Validate published results against current README
