+++ .github/workflows/ci-validate-schema.yml
@@ -22,19 +22,6 @@ jobs:
       - name: Install dependencies
         run: pip install -e .
 
-      - name: Check dataset_features.yml exists
-        run: |
-          if [ ! -e src/agenteval/leaderboard/dataset_features.yml ]; then
-            echo "dataset_features.yml is missing. Please run 'python scripts/update_schema.py' and commit the file to the repo." >&2
-            exit 1
-          fi
+      - name: Check if schema is up to date
+        run: python scripts/update_schema.py --check
 
-      - name: Regenerate dataset_features.yml
-        run: python scripts/update_schema.py
-
-      - name: Verify schema is up to date
-        run: |
-          if ! git diff --quiet src/agenteval/leaderboard/dataset_features.yml; then
-            echo "dataset_features.yml is out of date. Please run 'python scripts/update_schema.py' and commit the updated file." >&2
-            exit 1
-          fi
