+++ src/agenteval/cli.py
@@ -5,16 +5,19 @@ import re
 import subprocess
 import sys
 import tempfile
+from collections import defaultdict
 from datetime import datetime, timezone
 from io import BytesIO
 
 import click
 import datasets
 
+from agenteval.leaderboard.schema_generator import load_dataset_features
+
 from .cli_utils import AliasedChoice, generate_choice_help
 from .config import load_suite_config
 from .io import atomic_write_file
-from .leaderboard.models import LeaderboardSubmission
+from .leaderboard.models import LeaderboardSubmission, Readme
 from .leaderboard.upload import (
     compress_model_usages,
     sanitize_path_component,
@@ -24,7 +27,7 @@ from .models import EvalConfig, SubmissionMetadata, TaskResults
 from .score import process_eval_logs
 from .summary import compute_summary_statistics
 
-HF_URL_PATTERN = r"^hf://(?P<repo_id>[^/]+/[^/]+)/(?P<path>.*)$"
+HF_URL_PATTERN = r"^hf://(?:datasets/)?(?P<repo_id>[^/]+/[^/]+)/(?P<path>.*)$"
 EVAL_CONFIG_FILENAME = "eval_config.json"
 SCORES_FILENAME = "scores.json"
 SUMMARY_FILENAME = "summary_stats.json"
@@ -533,6 +536,9 @@ def publish_lb_command(repo_id: str, submission_urls: tuple[str, ...]):
         )
 
         # Create results files locally
+        config_splits = defaultdict(
+            list
+        )  # Accumulate config names and splits being published
         for (
             submission_url,
             submission_path,
@@ -573,6 +579,7 @@ def publish_lb_command(repo_id: str, submission_urls: tuple[str, ...]):
             eval_config = EvalConfig.model_validate_json(
                 open(local_eval_config_path).read()
             )
+            config_splits[eval_config.suite_config.name].append(eval_config.split)
             results = TaskResults.model_validate_json(
                 open(local_scores_path).read()
             ).results
@@ -597,6 +604,37 @@ def publish_lb_command(repo_id: str, submission_urls: tuple[str, ...]):
                 encoding="utf-8",
             ) as f:
                 f.write(lb_submission.model_dump_json(indent=None))
+
+        # Validate the config with the schema in HF
+        readme = Readme.download_and_parse(repo_id)
+        missing_configs = list(set(config_splits.keys()) - set(readme.configs.keys()))
+        if missing_configs:
+            click.echo(
+                f"Config name {missing_configs} not present in hf://{repo_id}/README.md"
+            )
+            click.echo(
+                f"Run 'update_readme.py add-config --repo-id {repo_id} --config-name {missing_configs[0]}' to add it"
+            )
+            sys.exit(1)
+        missing_splits = list(
+            set(((c, s) for c in config_splits.keys() for s in config_splits[c]))
+            - set(((c, s) for c in readme.configs.keys() for s in readme.configs[c]))
+        )
+        if missing_splits:
+            click.echo(
+                f"Config/Split {missing_splits} not present in hf://{repo_id}/README.md"
+            )
+            click.echo(
+                f"Run 'update_readme.py add-config --repo-id {repo_id} --config-name {missing_splits[0][0]} --split {missing_splits[0][1]}` to add it"
+            )
+            sys.exit(1)
+        local_features = load_dataset_features()
+        if local_features.arrow_schema != readme.features.arrow_schema:
+            click.echo(
+                "Schema in local dataset_features.yml does not match schema in hf://{repo_id}/README.md"
+            )
+            click.echo("Run 'update_readme.py sync-schema' to update it")
+            sys.exit(1)
 
         # Upload all results files in one shot
         click.echo(f"Uploading {len(submission_paths)} results to {repo_id}...")
